refactor: rearrange test folder
